+++ dom/bluetooth/bluez/BluetoothDBusService.h
@@ -207,19 +207,19 @@ public:
                       BluetoothReplyRunnable* aRunnable) override;
 
   virtual void
-  ReplyToMapFolderListing(long aMasId, const nsAString& aFolderlists,
+  ReplyToMapFolderListing(uint8_t aMasId, const nsAString& aFolderlists,
                           BluetoothReplyRunnable* aRunnable) override;
 
   virtual void
   ReplyToMapMessagesListing(BlobParent* aBlobParent, BlobChild* aBlobChild,
-                            long aMasId,
+                            uint8_t aMasId,
                             bool aNewMessage,
                             const nsAString& aTimestamp,
                             int aSize,
                             BluetoothReplyRunnable* aRunnable) override;
 
   virtual void
-  ReplyToMapMessagesListing(long aMasId,
+  ReplyToMapMessagesListing(uint8_t aMasId,
                             Blob* aBlob,
                             bool aNewMessage,
                             const nsAString& aTimestamp,
@@ -229,27 +229,27 @@ public:
   virtual void
   ReplyToMapGetMessage(BlobParent* aBlobParent,
                        BlobChild* aBlobChild,
-                       long aMasId,
+                       uint8_t aMasId,
                        BluetoothReplyRunnable* aRunnable) override;
 
   virtual void
   ReplyToMapGetMessage(Blob* aBlob,
-                       long aMasId,
+                       uint8_t aMasId,
                        BluetoothReplyRunnable* aRunnable) override;
 
   virtual void
-  ReplyToMapSetMessageStatus(long aMasId,
+  ReplyToMapSetMessageStatus(uint8_t aMasId,
                              bool aStatus,
                              BluetoothReplyRunnable* aRunnable) override;
 
   virtual void
-  ReplyToMapSendMessage(long aMasId,
+  ReplyToMapSendMessage(uint8_t aMasId,
                         const nsAString& aHandleId,
                         bool aStatus,
                         BluetoothReplyRunnable* aRunnable) override;
 
   virtual void
-  ReplyToMapMessageUpdate(long aMasId,
+  ReplyToMapMessageUpdate(uint8_t aMasId,
                           bool aStatus,
                           BluetoothReplyRunnable* aRunnable) override;
 
@@ -293,13 +293,13 @@ public:
                    const nsAString& aMessage) override;
 
   virtual void
-  SendMessageEvent(long aMasId,
+  SendMessageEvent(uint8_t aMasId,
                    BlobParent* aBlobParent,
                    BlobChild* aBlobChild,
                    BluetoothReplyRunnable* aRunnable) MOZ_OVERRIDE;
 
   virtual void
-  SendMessageEvent(long aMasId,
+  SendMessageEvent(uint8_t aMasId,
                    nsIDOMBlob* aBlob,
                    BluetoothReplyRunnable* aRunnable) MOZ_OVERRIDE;
