+++ dom/bluetooth/ipc/BluetoothServiceChildProcess.h
@@ -214,21 +214,21 @@ public:
                       BluetoothReplyRunnable* aRunnable) override;
 
   virtual void
-  ReplyToMapFolderListing(long aMasId,
+  ReplyToMapFolderListing(uint8_t aMasId,
                           const nsAString& aFolderlists,
                           BluetoothReplyRunnable* aRunnable) override;
 
   virtual void
   ReplyToMapMessagesListing(BlobParent* aBlobParent,
                             BlobChild* aBlobChild,
-                            long aMasId,
+                            uint8_t aMasId,
                             bool aNewMessage,
                             const nsAString& aTimestamp,
                             int aSize,
                             BluetoothReplyRunnable* aRunnable) override;
 
   virtual void
-  ReplyToMapMessagesListing(long aMasId,
+  ReplyToMapMessagesListing(uint8_t aMasId,
                             Blob* aBlob,
                             bool aNewMessage,
                             const nsAString& aTimestamp,
@@ -238,28 +238,28 @@ public:
   virtual void
   ReplyToMapGetMessage(BlobParent* aBlobParent,
                        BlobChild* aBlobChild,
-                       long aMasId,
+                       uint8_t aMasId,
                        BluetoothReplyRunnable* aRunnable) override;
 
   virtual void
   ReplyToMapGetMessage(Blob* aBlob,
-                       long aMasId,
+                       uint8_t aMasId,
                        BluetoothReplyRunnable* aRunnable) override;
 
   virtual void
-  ReplyToMapSetMessageStatus(long aMasId,
+  ReplyToMapSetMessageStatus(uint8_t aMasId,
                              bool aStatus,
                              BluetoothReplyRunnable* aRunnable) override;
 
   virtual void
-  ReplyToMapSendMessage(long aMasId,
+  ReplyToMapSendMessage(uint8_t aMasId,
                         const nsAString& aHandleId,
                         bool aStatus,
                         BluetoothReplyRunnable* aRunnable) override;
 
   virtual void
   ReplyToMapMessageUpdate(
-    long aMasId, bool aStatus, BluetoothReplyRunnable* aRunnable) override;
+    uint8_t aMasId, bool aStatus, BluetoothReplyRunnable* aRunnable) override;
 
 #ifdef MOZ_B2G_RIL
   virtual void
